delete subcategories hook
